Add files via upload
You can download the dataset from Kaggle through this link https://www.kaggle.com/datasets/utkarshx27/heart-disease-diagnosis-dataset/data

IMP NOTE- ---Make sure you download and store the dataset in the same folder as where the code file is saved. ---Rename the dataset name as it is written in the code.